change kafka network settings (again)
Up until now, external kafka clients have been able to use our kafka broker,
but internal clients fail (on some clouds, internal clients cant see the public
ip that we advertised).

We had a hack to force entries into /etc/hosts on the broker and set the
hostname so at least the kakfa unit could write to its own topic. This worked,
but required us to expose kakfa even if we had no intention of using an
external client.

It turns out all you need to do is set the advertised.host.name to any name
that units in a provider can access (luckily, 'hostname -s' seems to do the
trick; tested azure/aws). Then instruct the admin via the readme that they
will need to add an /etc/hosts entry on their client to use our broker.

Yes, it is more work for the admin, but it is more maintainable than the
/etc/hosts and hostname hack, and does not expose kafka unnecessarily.